Sprint 12. Исправление ПР
