+++ app/src/main/java/com/example/playlistmaker/StateType.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker
+
+enum class StateType {
+    CONNECTION_ERROR,
+    NOT_FOUND,
+    SEARCH_RESULT,
+    HISTORY_LIST,
+}
