+++ app/src/main/java/com/example/playlistmaker/Track.kt
@@ -1,6 +1,5 @@
 package com.example.playlistmaker
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
+
 data class Track(
     val trackId: Int,
     val trackName: String,
@@ -8,25 +7,3 @@ data class Track(
     val trackTimeMillis: Int,
     val artworkUrl100: String
 )
-class TracksAdapter(
-    private val clickListener: ClickListener,
-) : RecyclerView.Adapter<TrackViewHolder>() {
-    var track : ArrayList<Track> = ArrayList()
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackViewHolder {
-        return TrackViewHolder(parent)
-
-    }
-
-    override fun onBindViewHolder(viewer: TrackViewHolder, position: Int) {
-        viewer.bind(track[position])
-
-        viewer.itemView.setOnClickListener{ clickListener.click(track[position])}
-    }
-
-    override fun getItemCount(): Int = track.size
-    fun interface ClickListener {
-        fun click(track: Track)
-    }
-
-}
