+++ app/src/main/java/com/example/playlistmaker/TrackViewHolder.kt
@@ -1,39 +1,14 @@
 package com.example.playlistmaker
-import android.util.Log;
 import android.content.res.Resources
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
-import android.widget.Toast
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.CenterCrop
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.bumptech.glide.request.RequestOptions
-import kotlin.collections.ArrayList
 
-
-class TrackAdapter(private val trackList: ArrayList<Track>) : RecyclerView.Adapter<TrackViewHolder>() {
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.activity_tracks, parent, false)
-        return TrackViewHolder(parent)
-    }
-    override fun onBindViewHolder(holder: TrackViewHolder, position: Int) {
-        val track = trackList[position]
-        holder.bind(track)
-    }
-    override fun getItemCount(): Int {
-        return trackList.size
-    }
-}
-
-
-//class TrackViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-
-    class TrackViewHolder(parentView: ViewGroup) : RecyclerView.ViewHolder(
+class TrackViewHolder(parentView: ViewGroup) : RecyclerView.ViewHolder(
         LayoutInflater.from(parentView.context).inflate(R.layout.activity_tracks, parentView, false)
     ) {
